FEAT (Chapter 8) : 문제에 맞게 Collatz로 변경
