+++ Chapter8/src/main/java/com/nhnacademy/Jminsoo/excercise2/Excercise8_2.java
@@ -7,7 +7,7 @@ public class Excercise8_2 {
      */
 
     public static void main(String[] args) {
-        BigIntegerApp app = new BigIntegerApp();
+        CollatzApp app = new CollatzApp();
         app.start();
     }
